gothello-ai/requirements.txt
asyncio==3.4.3
future==0.18.2
neat-python==0.92
numpy
progress==1.5
requests==2.22.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
websockets==8.1
click
